% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeLULCvsLSTrelationship.R
\name{AnalyzeLULCvsLSTrelationship}
\alias{AnalyzeLULCvsLSTrelationship}
\title{Analyze the Relationship Between LULC Change and LST Difference}
\usage{
AnalyzeLULCvsLSTrelationship(
  lulc_change_path,
  lst_difference_path,
  output_csv_path,
  plot_relationship = TRUE
)
}
\arguments{
\item{lulc_change_path}{A string specifying the file path of the LULC change raster map.}

\item{lst_difference_path}{A string specifying the file path of the LST difference raster map.}

\item{output_csv_path}{A string specifying the file path where the combined data should be saved as CSV.}

\item{plot_relationship}{A logical value indicating whether to plot the relationship between LULC change and LST difference. Default is TRUE.}
}
\value{
A list containing the correlation value, linear model summary, and the combined data.
}
\description{
This function analyzes the relationship between land use/land cover (LULC) change
and land surface temperature (LST) difference by extracting values from LULC and LST
raster datasets, performing correlation and regression analysis, and visualizing the relationship.
}
\examples{
lulc_lst_analysis <- AnalyzeLULCvsLSTrelationship(
  lulc_change_path = "E:/EAGLE/R_programming/LULC_LST/Change_Detection/ChangeDetection.tif",
  lst_difference_path = "E:/EAGLE/R_programming/LULC_LST/LST_Difference/LSTDifference.tif",
  output_csv_path = "E:/EAGLE/R_programming/LULC_LST/LULCvsLST/combined_data.csv",
  plot_relationship = TRUE
)
print(lulc_lst_analysis)
}
