% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeDetectionAnalysis.R
\name{ChangeDetectionAnalysis}
\alias{ChangeDetectionAnalysis}
\title{Change Detection Analysis and Area Calculation}
\usage{
ChangeDetectionAnalysis(
  classified_2014_path,
  classified_2024_path,
  plot_area_change = TRUE,
  save_plot = FALSE,
  plot_filename = "change_detection_map.png"
)
}
\arguments{
\item{classified_2014_path}{A string specifying the file path of the classified
raster map for the year 2014.}

\item{classified_2024_path}{A string specifying the file path of the classified
raster map for the year 2024.}

\item{plot_area_change}{A logical value indicating whether to plot the area change
by class. Default is TRUE.}

\item{save_plot}{A logical value indicating whether to save the generated change
detection map plot as a PNG file. Default is FALSE.}

\item{plot_filename}{A string specifying the file name to save the plot, if
\code{save_plot} is TRUE. Default is "change_detection_map.png".}
}
\value{
A data frame with the area (in square kilometers) for each change class
detected in the analysis.
}
\description{
This function performs change detection analysis between two classified raster maps
(for the years 2014 and 2024), generates a change detection map, calculates the
area change for each class, and optionally plots the results.
}
\examples{
change_area_results <- ChangeDetectionAnalysis(
  classified_2014_path="E:/EAGLE/R_programming/LULC_LST/Data_2014/Masked_Asmara_Classification_2014.tif",
  classified_2024_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/Masked_Asmara_Classification_2024_discrete.tif",
  plot_area_change = TRUE,
  save_plot = TRUE,
  plot_filename = "change_detection_map.png"
)
}
