% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyLanduseLandcover.R
\name{ClassifyLanduseLandcover}
\alias{ClassifyLanduseLandcover}
\title{Land Use Land Cover Classification Using Random Forest}
\usage{
ClassifyLanduseLandcover(raster_path, shapefile_path, output_raster_path)
}
\arguments{
\item{raster_path}{A character string specifying the path to the input raster file.}

\item{shapefile_path}{A character string specifying the path to the input shapefile.}

\item{output_raster_path}{A character string specifying the path to save the output classified raster.}
}
\value{
A raster object containing the classified land use/land cover classes.
}
\description{
This function classifies a land use/land cover raster using Random Forest (RF)
based on provided training data. It reads a raster and shapefile of training
data, extracts spectral values from the raster at the centroids of polygons
in the shapefile, and trains a Random Forest model to predict the classes
across the raster. The function then saves the classified raster and plots it.
}
\examples{
classified_raster <- ClassifyLanduseLandcover(
  raster_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/MaskedAsmara20214.tif",
  shapefile_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Training_2014/Training_2014.shp",
  output_raster_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Masked_Asmara_Classification_2014.tif"
)
}
