% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrix.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{Confusion Matrix Function for Validation of Classified Raster}
\usage{
ConfusionMatrix(validation_shapefile_path, classified_raster_path)
}
\arguments{
\item{validation_shapefile_path}{A character string representing the path
to the shapefile containing the ground truth validation data. This shapefile
should have polygon features with an associated class label (id).}

\item{classified_raster_path}{A character string representing the path to
the classified raster file. This raster should contain pixel values representing
the predicted class for each pixel.}
}
\value{
A confusion matrix object with various metrics, including accuracy and Kappa.
The confusion matrix is printed to the console, and accuracy and Kappa values are also displayed.
}
\description{
This function computes a confusion matrix by comparing the predicted values
from a classified raster with the actual values from a validation shapefile.
It also computes the overall accuracy and Kappa coefficient based on the
confusion matrix.
}
\examples{
# Example usage of the function
ConfusionMatrix(
  validation_shapefile_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Validation_2014/Validation_2014/Validation_2014.shp",
  classified_raster_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Masked_Asmara_Classification_2014.tif"
)

}
