% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrixfirst.R
\name{ConfusionMatrixfirst}
\alias{ConfusionMatrixfirst}
\title{Confusion Matrix Calculation for Classified Raster}
\usage{
ConfusionMatrixfirst(
  raster_path,
  validation_shapefile_path,
  classified_raster_path
)
}
\arguments{
\item{raster_path}{Character string. The file path to the raster that will
be used for analysis (the masked raster).}

\item{validation_shapefile_path}{Character string. The file path to the shapefile
containing the ground truth validation data for classification.}

\item{classified_raster_path}{Character string. The file path to the classified raster
that contains the predicted classification values.}
}
\value{
A confusion matrix, including overall accuracy and Kappa coefficient.
If the data is incompatible, an error message is printed.
}
\description{
This function calculates the confusion matrix for the classification results
of a raster dataset, using validation data from a shapefile. The function
compares the predicted classification from a raster with the ground truth
data (validation shapefile) and calculates overall accuracy and Kappa coefficient.
}
\examples{
ConfusionMatrixfirst(
  raster_path = "E:/path/to/masked_Asmara.tif",
  validation_shapefile_path = "E:/path/to/Validation_2024.shp",
  classified_raster_path = "E:/path/to/Masked_Asmara_Classification_2024_discrete.tif"
)
}
