% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{LinearRegression}
\alias{LinearRegression}
\title{Perform Linear Regression between LST and NDVI (for 2014)}
\usage{
LinearRegression(LST_path, NDVI_path, output_plot_path = NULL)
}
\arguments{
\item{LST_path}{A character string representing the file path to the LST raster file for 2014.}

\item{NDVI_path}{A character string representing the file path to the NDVI raster file for 2014.}

\item{output_plot_path}{A character string representing the file path where the plot should
be saved. If \code{NULL}, the plot will not be saved. (default is \code{NULL}).}
}
\value{
A list with two elements:
\item{plot}{A ggplot object representing the scatter plot with the regression line.}
\item{lm_model}{The linear model object generated by \code{lm()} containing the results of
the regression analysis.}
}
\description{
This function performs a linear regression analysis between Land Surface Temperature (LST)
and Normalized Difference Vegetation Index (NDVI) for the year 2014. It also generates a scatter plot with
the linear regression line and optionally saves the plot to a file.
}
\examples{
result <- LinearRegression(
  LST_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/LST_2014/LandSurfaceTemeprature_2014.tif",
  NDVI_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/LST_2014/NDVI_2014.tif",
  output_plot_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/LST_NDVI_regression_plot_2014.png"
)
summary(result$lm_model)
print(result$plot)

}
