% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegressionfirst.R
\name{LinearRegressionfirst}
\alias{LinearRegressionfirst}
\title{Perform Linear Regression between LST and NDVI}
\usage{
LinearRegressionfirst(LST_path, NDVI_path, output_plot_path = NULL)
}
\arguments{
\item{LST_path}{A character string representing the file path to the LST raster file.}

\item{NDVI_path}{A character string representing the file path to the NDVI raster file.}

\item{output_plot_path}{A character string representing the file path where the plot should
be saved. If \code{NULL}, the plot will not be saved. (default is \code{NULL}).}
}
\value{
A list with two elements:
\item{plot}{A ggplot object representing the scatter plot with the regression line.}
\item{lm_model}{The linear model object generated by \code{lm()} containing the results of
the regression analysis.}
}
\description{
This function performs a linear regression analysis between Land Surface Temperature (LST)
and Normalized Difference Vegetation Index (NDVI). It also generates a scatter plot with
the linear regression line and optionally saves the plot to a file.
}
\examples{
result <- LinearRegressionfirst(
  LST_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/LST_2024/LandSurfaceTemeprature_2024.tif",
  NDVI_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/LST_2024/NDVI_2024.tif",
  output_plot_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/LST_NDVI_regression_plot.png"
)
summary(result$lm_model)
print(result$plot)

}
