% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LST_DifferenceAnalysis.R
\name{LST_DifferenceAnalysis}
\alias{LST_DifferenceAnalysis}
\title{LST Difference Analysis and Visualization}
\usage{
LST_DifferenceAnalysis(
  lst_2014_path,
  lst_2024_path,
  asmara_shapefile_path,
  threshold = 0.5,
  plot_diff = TRUE,
  plot_significant_diff = TRUE,
  save_plot = FALSE,
  plot_filename_diff = "lst_diff_map.png",
  plot_filename_significant_diff = "lst_significant_diff_map.png"
)
}
\arguments{
\item{lst_2014_path}{A string specifying the file path of the LST raster for the year 2014.}

\item{lst_2024_path}{A string specifying the file path of the LST raster for the year 2024.}

\item{asmara_shapefile_path}{A string specifying the file path of the Asmara shapefile.}

\item{threshold}{A numeric value for the threshold to identify significant differences.
Differences below this threshold are considered similar and set to NA.}

\item{plot_diff}{A logical value indicating whether to plot the LST difference. Default is TRUE.}

\item{plot_significant_diff}{A logical value indicating whether to plot significant LST differences. Default is TRUE.}

\item{save_plot}{A logical value indicating whether to save the plots as PNG files. Default is FALSE.}

\item{plot_filename_diff}{A string specifying the file name to save the LST difference plot if \code{save_plot} is TRUE.}

\item{plot_filename_significant_diff}{A string specifying the file name to save the significant LST difference plot if \code{save_plot} is TRUE.}
}
\value{
A list containing summary statistics for the LST difference, along with the data frames for the LST differences and significant differences.
}
\description{
This function performs analysis on the Land Surface Temperature (LST) difference
between two years (2014 and 2024), visualizes the difference, identifies significant
differences, and calculates summary statistics.
}
\examples{
lst_analysis_results <- LST_DifferenceAnalysis(
  lst_2014_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/LST_2014/LandSurfaceTemeprature_2014.tif",
  lst_2024_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/LST_2024/LandSurfaceTemeprature_2024.tif",
  asmara_shapefile_path = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Greater_Asmara_Shapefile/Asmara.shp",
  threshold = 0.5,
  plot_diff = TRUE,
  plot_significant_diff = TRUE,
  save_plot = TRUE,
  plot_filename_diff = "lst_diff_map.png",
  plot_filename_significant_diff = "lst_significant_diff_map.png"
)
}
