% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LULCvsLST.R
\name{LULCvsLST}
\alias{LULCvsLST}
\title{LULC vs LST Analysis Function for 2014}
\usage{
LULCvsLST(LST_file, Classified_file, output_dir = NULL)
}
\arguments{
\item{LST_file}{Path to the LST raster file (in .tif format).}

\item{Classified_file}{Path to the classified raster file (in .tif format).}

\item{output_dir}{Optional parameter specifying the output directory to set as
the working directory. If not provided, the current working directory remains unchanged.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{class_means}: Data frame with the mean LST for each land cover class.
\item \code{anova}: Summary of the ANOVA test results.
\item \code{lm_summary}: Summary of the linear regression model.
}
}
\description{
This function performs a series of analyses comparing Land Surface Temperature (LST)
and Land Use/Land Cover (LULC) classes. It includes resampling the LST raster,
extracting values from the rasters, calculating mean LST by class, performing an
ANOVA test for differences between classes, and fitting a linear regression model.
It also creates a boxplot of LST by class.
}
\examples{
LULCvsLST(LST_file = "E:/EAGLE/R_programming/LULC_LST/Data_2014/LST_2014/LandSurfaceTemeprature_2014.tif",
               Classified_file = "E:/EAGLE/R_programming/LULC_LST/Data_2014/Masked_Asmara_Classification_2014.tif",
               output_dir = "E:/EAGLE/R_programming/LULC_LST/Data_2014")
}
