% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LULCvsLSTfirst.R
\name{LULCvsLSTfirst}
\alias{LULCvsLSTfirst}
\title{LULC vs LST Analysis Function}
\usage{
LULCvsLSTfirst(LST_file, Classified_file, output_dir = NULL)
}
\arguments{
\item{LST_file}{Path to the LST raster file (in .tif format).}

\item{Classified_file}{Path to the classified raster file (in .tif format).}

\item{output_dir}{Optional parameter specifying the output directory to set as
the working directory. If not provided, the current working directory remains unchanged.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{class_means}: Data frame with the mean LST for each land cover class.
\item \code{anova}: Summary of the ANOVA test results.
\item \code{lm_summary}: Summary of the linear regression model.
}
}
\description{
This function performs a series of analyses comparing Land Surface Temperature (LST)
and Land Use/Land Cover (LULC) classes. It includes resampling the LST raster,
extracting values from the rasters, calculating mean LST by class, performing an
ANOVA test for differences between classes, and fitting a linear regression model.
It also creates a boxplot of LST by class.
}
\examples{
LULCvsLSTfirst(LST_file = "path/to/LST.tif",
               Classified_file = "path/to/Classified.tif",
               output_dir = "path/to/output/dir")
}
