% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingLULCfirst.R
\name{PlottingLULCfirst}
\alias{PlottingLULCfirst}
\title{Title}
\usage{
PlottingLULCfirst(classified_raster_path, pixel_area_km2 = 9e-04)
}
\arguments{
\item{classified_raster_path}{path to the classified raster data (e.g., a GeoTIFF file)}

\item{pixel_area_km2}{Area of a single pixel in square kilometers. Default is 0.0009km2}
}
\value{
A data frame conatining the area and percenatge of each Land use/ Land cover class
}
\description{
Title
}
\examples{
\dontrun{
classified_raster_path <- "E:/EAGLE/R_programming/LULC_LST/Data_2024/Masked_Asmara_Classification_2024_discrete.tif"
PlottingLULCfirst(classified_raster_path)
}
}
