% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingLandsat.R
\name{ProcessingLandsat}
\alias{ProcessingLandsat}
\title{Process Landsat Data for 2014}
\usage{
ProcessingLandsat(
  landsat_dir,
  shapefile_path,
  output_path,
  bands = c(1, 2, 3, 4, 5, 7)
)
}
\arguments{
\item{landsat_dir}{The directory where the Landsat band files for 2014 are stored.}

\item{shapefile_path}{The path to the shapefile used for cropping the Landsat image.}

\item{output_path}{The path where the masked raster image will be saved.}

\item{bands}{A vector of band numbers to be used from the Landsat image. Default is c(1, 2, 3, 4, 5, 7).}
}
\value{
A masked raster image of the Landsat data.
}
\description{
This function reads, processes, and crops Landsat 8 images for 2014 based on the provided bands,
and then masks the image with a given shapefile. It returns the processed raster image.
}
\examples{
landsat_dir <- "E:/EAGLE/R_programming/LULC_LST/Data_2014/LC08_L2SP_169049_20140125_20200912_02_T1"
shapefile_path <- "E:/EAGLE/R_programming/LULC_LST/Data_2014/Greater_Asmara_Shapefile/Asmara.shp"
output_path <- "E:/EAGLE/R_programming/LULC_LST/Data_2014/MaskedAsmara20214.tif"
masked_asmara_image <- ProcessingLandsat(landsat_dir, shapefile_path, output_path)
}
