% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingLandsat2024.R
\name{ProcessingLandsat2024}
\alias{ProcessingLandsat2024}
\title{Process Landsat Data for 2024}
\usage{
ProcessingLandsat2024(
  landsat_dir,
  shapefile_path,
  output_path,
  bands = c(1, 2, 3, 4, 5, 7),
  plot_processing = TRUE,
  save_plot = FALSE,
  plot_filename = "landsat_processing_plot.png"
)
}
\arguments{
\item{landsat_dir}{The directory where the Landsat band files are stored.}

\item{shapefile_path}{The path to the shapefile used for cropping the Landsat image.}

\item{output_path}{The path where the masked raster image will be saved.}

\item{bands}{A vector of band numbers to be used from the Landsat image. Default is c(1, 2, 3, 4, 5, 7).}

\item{plot_processing}{A logical value indicating whether to plot the processed images. Default is TRUE.}

\item{save_plot}{A logical value indicating whether to save the generated plot as a PNG file. Default is FALSE.}

\item{plot_filename}{A string specifying the file name to save the plot, if save_plot is TRUE. Default is "landsat_processing_plot.png".}
}
\value{
A masked raster image of the Landsat data.
}
\description{
This function reads, processes, and crops Landsat 8 images based on the provided bands,
and then masks the image with a given shapefile. It returns the processed raster image.
}
\examples{
landsat_dir <- "E:/EAGLE/R_programming/LULC_LST/Data_2024/LC08_L2SP_169049_20240309_20240316_02_T1"
shapefile_path <- "E:/EAGLE/R_programming/LULC_LST/Data_2024/Greater_Asmara_Shapefile/Asmara.shp"
output_path <- "E:/EAGLE/R_programming/LULC_LST/Data_2024/Masked_Asmara_2024.tif"
masked_asmara_image <- ProcessingLandsat2024(landsat_dir, shapefile_path, output_path)
}
