% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingLandsat2024.R
\name{ProcessingLandsat2024}
\alias{ProcessingLandsat2024}
\title{Processing Landsat 8 with clipping and masking for 2024}
\usage{
ProcessingLandsat2024(
  landsat_dir,
  shapefile_path,
  output_path,
  bands = c(1, 2, 3, 4, 5, 7)
)
}
\arguments{
\item{landsat_dir}{A character string representing the directory path where the Landsat data files are stored.}

\item{shapefile_path}{A character string representing the file path to the shapefile containing the boundary (e.g., administrative area) used for clipping and masking the Landsat image.}

\item{output_path}{A character string specifying the path where the output masked Landsat image will be saved.}

\item{bands}{A numeric vector specifying the Landsat bands to be processed. Defaults to bands 1, 2, 3, 4, 5, and 7, which correspond to the blue, green, red, NIR, SWIR1, and SWIR2 bands.}
}
\value{
A SpatRaster object (from the \code{terra} package) representing the masked Landsat image after clipping to the provided shapefile's boundary.
}
\description{
This function processes Landsat 8 surface reflectance data by clipping and masking
it according to a provided shapefile boundary. The function accepts a directory
of Landsat data, a shapefile for clipping, and outputs a masked Landsat image.
}
\examples{
# Example usage:
ProcessingLandsat2024(
  landsat_dir = "E:/EAGLE/R_programming/LULC_LST/Data_2024/LC08_L2SP_169049_20240309_20240316_02_T1/",
  shapefile_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/Greater_Asmara_Shapefile/Asmara.shp",
  output_path = "E:/EAGLE/R_programming/LULC_LST/Data_2024/Masked_Asmara_2024.tif",
  bands = c(1, 2, 3, 4, 5, 7)
)
}
