% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessingLandsat2024.R
\name{ProcessingLandsat2024}
\alias{ProcessingLandsat2024}
\title{Process Landsat Data for 2024}
\usage{
ProcessingLandsat2024(
  landsat_dir,
  shapefile_path,
  output_path,
  bands = c(1, 2, 3, 4, 5, 7)
)
}
\arguments{
\item{landsat_dir}{The directory where the Landsat band files are stored.}

\item{shapefile_path}{The path to the shapefile used for cropping the Landsat image.}

\item{output_path}{The path where the masked raster image will be saved.}

\item{bands}{A vector of band numbers to be used from the Landsat image. Default is c(1, 2, 3, 4, 5, 7).}
}
\value{
A masked raster image of the Landsat data.
}
\description{
This function reads, processes, and crops Landsat 8 images based on the provided bands,
and then masks the image with a given shapefile. It returns the processed raster image.
}
